This commit was manufactured by cvs2svn to create tag 'Root_SpecTcl-2-2_development'.

+++ SpecTcl/Display/dispshare.h
@@ -341,8 +341,8 @@ typedef union {
 	      } spec_spectra; /* Spectrum storage type. */
 
 typedef struct {
-                 unsigned short xchans;
-		 unsigned short ychans;
+                 unsigned int xchans;
+		 unsigned int ychans;
 	       } spec_dimension;   /* Describes the channels in a spectrum. */
 
 typedef char spec_title[72];
